requirements.txt
flask
flask-cors
pymongo
dnspython
python-dotenv
